pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "openclaw-memory"
version = "1.0.0"
description = "AI Agent Long-term Memory System with Vector Storage & Knowledge Graph"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "OpenClaw Team"}
]
keywords = ["memory", "vector-database", "knowledge-graph", "ai-agent", "weaviate", "ollama"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Artificial Intelligence",
]

dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "weaviate-client>=4.9.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "watchdog>=3.0.0",
    "schedule>=1.2.0",
    "tqdm>=4.65.0",
    "pydantic>=2.0.0",
    "Pillow>=10.0.0",
    "PyPDF2>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
]

[project.optional-dependencies]
nebula = ["nebula3-python>=3.6.0"]
kg = ["networkx>=3.0"]
embedding = ["sentence-transformers>=2.7.0"]
all = ["openclaw-memory[nebula,kg,embedding]"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
openclaw-memory = "main:main"

[tool.setuptools.packages.find]
include = ["openclaw_memory*"]

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
